% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the cyclestatsv3 Shiny Application}
\usage{
run_app(
  on_start = NULL,
  options = list(),
  enable_bookmarking = NULL,
  ui_pattern = "/",
  ...
)
}
\arguments{
\item{on_start}{Passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}

\item{options}{Named list of options passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}

\item{enable_bookmarking}{Passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}

\item{ui_pattern}{Passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}

\item{...}{Additional arguments passed as golem options.}
}
\description{
Main entry point. Initialises data then launches the app.
}
